requirements.txt
Pillow == 12.1.1